Created types and function pages to separate from schedule.tsx - adding in section/schedule functions for backend calls
